Moved event namespaced features into event namespaces
