+++ src/resources/namespaced/event-interests.ts
@@ -1,33 +1,34 @@
-import {createRequestMethod} from "../interfaces/server/request-method.interface";
-import {ResourceInterface} from "../interfaces/server/resource";
+import {createRequestMethod} from "../../interfaces/server/request-method.interface";
+import {ResourceInterface} from "../../interfaces/server/resource";
 
 export default {
-    path: "events/{event_id}/interest",
 
     methods: [
         /**
          * Add your interest to an event
-         * @name create
+         * @name interests.create
          * @return EventInterestInterface
          */
         createRequestMethod({
+            namespace:'interests',
             name: "create",
             method: "POST",
-            path: "",
-            required: ["event_id"],
+            path: "/{id}/interest",
+            required: ["id"],
             requiresAuth: true
         }),
 
         /**
          * Remove your interest from an event
-         * @name delete
+         * @name interests.delete
          * @return EventInterestInterface
          */
         createRequestMethod({
+            namespace:'interests',
             name: "delete",
             method: "DELETE",
-            path: "",
-            required: ["event_id"],
+            path: "/{id}/interest",
+            required: ["id"],
             requiresAuth: true
         }),
